+++ NEWS.md
@@ -1,3 +1,31 @@
+# peskas.timor.data.pipeline 0.16.0
+
+## Breaking changes
+
+- The Docker file now run rocker/r-ver 4.1.1 instead of rocker/geospatial:4.0.3 
+
+## Bug fixes
+
+- Fixed the bug when calling `ingest_rfish_table()` in the main pipeline.
+
+
+# peskas.timor.data.pipeline 0.15.0
+
+## New features
+
+- Added a series of functions (`get_catch_types()`, `get_fish_length()` ,`retrieve_lengths()`) useful to retrieve morphometric conversion factors from catch types names in metadata tables. 
+- Added a new metadata table in airtable (morphometric_table) containing length-weight and length-length conversion factors.
+- Added `join_weights()` which integrates morphometric data with merged landings.
+
+
+# peskas.timor.data.pipeline 0.14.0
+
+## New features
+
+- We test validated data to ensure its integrity
+- Added `merge_trips()` which integrates data from the landings and tracking together
+- Added `format_public_data()` to format and export data to be used for analytics
+
 # peskas.timor.data.pipeline 0.13.0
 
 ## New features